+++ server/api/blog.js
@@ -8,13 +8,13 @@ export default defineEventHandler(async (event) => {
   const { slug } = getQuery(event);
 
   if (slug) {
-    const { data } = await client.from('pages')
+    const { data } = await client.from('blog')
       .select('*')
       .eq('slug', slug)
       .single();
     return data;
   } else {
-    const { data } = await client.from('pages')
+    const { data } = await client.from('blog')
       .select('*');
     return data;
   }